update to all code
Lots changed, but the overview is that I took all the functions we made (data cleaning/processing, making dictionaries, classifier) and put them in the code, create 5 sets of training/testing data and ran the code on these 5 sets of data, which outputs the accuracy. Still to do: create function for user input, visuals & ensure everything is annotated.